Add chat delay
Adds a race option to delay chat. Users who are able to monitor races are not affected and see the messages immediately. The user who sends the message also sees the message immediately.

Known issues: Adding or removing monitor requires a refresh before the change takes affect for the chat delay. Likely the same if a user becomes a moderator while already on a race page.