Update view to enable support for category
While category was missing as a flag, I think we don't need to add in finish_time and user info as its own for just the requested user. If they want the rest of that I think it would be best if they pass show_entrants to get that and parse from that.
Willing to be moved on that.

For the userdata, do you have an efficent way to pull the number of times a user has raced a goal/personal time per the issue?
As is I've appended goals to the category response in general as it seems like useful information.