CSS update to reduce race room squishing.
Setting a min-width on the main tag ensures that the the main content area is at minimum double the with of the aside (chat) area.  This is a band-aid on a larger problem that is caused by the flex-basis on the aside element, which causes the main area to grow irregularly. However, just removing the flex-basis on the aside element makes the side-effect of chat causing flex recalculations on the main container much more apparent, with resizing happening much more frequently. A restructure of the css will most likely be required for a more fluid look.

I also removed a mostly useless media query, which conflicted with stated design intentions (race containers shouldn't exceed 1280px). I added a media query that allows for fluid shrinking of the main content still, keeping expected functionality consistent.